# SpandaServer/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ITodo": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"description": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"description"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "spandaserver",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/spanda/users": {
			"get": {
				"operationId": "IsUserAuthenticated",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "Register",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"id": {
									"type": "string"
								},
								"password": {
									"type": "string"
								},
								"email": {
									"type": "string"
								},
								"phone": {
									"type": "string"
								},
								"isAutoUpdateEnabled": {
									"type": "boolean"
								}
							},
							"title": "RegisterBody",
							"type": "object",
							"required": [
								"id",
								"password",
								"email",
								"phone",
								"isAutoUpdateEnabled"
							]
						}
					}
				]
			}
		},
		"/spanda/oauth/token": {
			"post": {
				"operationId": "SetNewRefreshAndAccessToken",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"refresh_token": {
									"type": "string"
								}
							},
							"title": "SetNewRefreshAndAccessTokenBody",
							"type": "object",
							"required": [
								"refresh_token"
							]
						}
					}
				]
			}
		},
		"/spanda/banks/{blz}": {
			"get": {
				"operationId": "GetBankByBLZ",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "blz",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/todo": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/ITodo"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"description": {
									"type": "string"
								}
							},
							"title": "CreateBody",
							"type": "object",
							"required": [
								"description"
							]
						}
					}
				]
			}
		},
		"/todo/{id}": {
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"description": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"description"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}